Move pipeline integration tests to higher level in tree
